+++ lib/constants.dart
@@ -15,3 +15,4 @@ const kCacheArrivals = Duration(seconds: 10);
 const kHostToPing = 'transport.tallinn.ee';
 const kArrivalsMessageStyle = TextStyle(fontSize: 20.0);
 const kRouteToSelfDistance = 1500; // meters
+const kStopsTooCloseDistance = 10; // meters