+++ lib/providers/arrivals.dart
@@ -9,7 +9,7 @@ import 'package:quick_bus/models/bus_stop.dart';
 
 final _arrivalsCache = ArrivalsCache();
 
-final arrivalsProvider = FutureProvider.autoDispose
+final arrivalsProvider = FutureProvider
     .family<List<Arrival>, BusStop?>((ref, stop) async {
   if (stop == null) return [];
   final cached = _arrivalsCache.find(stop);
@@ -19,10 +19,8 @@ final arrivalsProvider = FutureProvider.autoDispose
   // print('Updating arrivals for $stopStr');
   List<Arrival> arrivals = const [];
   try {
-    if (stop is SiriBusStop)
-      arrivals = await SiriHelper().getArrivals(stop);
-    if (arrivals.isEmpty)
-      arrivals = await RouteQuery().getArrivals(stop);
+    if (stop is SiriBusStop) arrivals = await SiriHelper().getArrivals(stop);
+    if (arrivals.isEmpty) arrivals = await RouteQuery().getArrivals(stop);
   } on SocketException catch (e) {
     // TODO: show dialog, but just one time.
     throw ArrivalFetchError(e.toString());
@@ -31,6 +29,25 @@ final arrivalsProvider = FutureProvider.autoDispose
   }
   _arrivalsCache.add(stop, arrivals);
   return arrivals;
+});
+
+final multipleArrivalsProvider = FutureProvider.autoDispose
+    .family<List<Arrival>, BusStop?>((ref, stop) async {
+  if (stop == null) return [];
+  List<BusStop> stops = [stop, ...stop.stopsAround];
+
+  List<Arrival> result = [];
+  String? errorMessage;
+  for (final curStop in stops) {
+    try {
+      result.addAll(await ref.read(arrivalsProvider(curStop).future));
+    } on ArrivalFetchError catch (e) {
+      errorMessage = e.message;
+    }
+  }
+  if (result.isEmpty && errorMessage != null)
+    throw ArrivalFetchError(errorMessage);
+  return result;
 });
 
 // Taking a hint from https://github.com/rrousselGit/river_pod/issues/461#issuecomment-825837140